todo

+++ src/inet/clock/base/DriftingOscillatorBase.h
@@ -24,7 +24,7 @@ class INET_API DriftingOscillatorBase : public OscillatorBase, public IScriptabl
     simtime_t nominalTickLength;
     ppm driftRate = ppm(NaN); // 0 means nominal, higher value means faster oscillator, e.g. 100 ppm means the oscillator gains 100 microseconds for every second in simulation time
                               // 100 ppm value means the current tick length is smaller by a factor of (1 / (1 + 100 / 1E+6)) than the nominal tick length measured in simulation time
-    long double driftFactor;
+    double driftFactor;
 
     simtime_t origin; // simulation time from which the computeClockTicksForInterval and computeIntervalForClockTicks is measured, it is always in the past
     simtime_t nextTickFromOrigin; // simulation time interval from the computation origin to the next tick
